# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the docket app",
        "title": "Docket API",
        "contact": {
            "name": "Nayan Das",
            "email": "dastms@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/oauth/token": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Oauth bearer token",
                "parameters": [
                    {
                        "description": "Create token",
                        "name": "grants",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.OauthTokenBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/controllers.OauthTokenResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.OauthTokenBody": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "string"
                },
                "client_secret": {
                    "type": "string"
                },
                "grant_type": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "controllers.OauthTokenData": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "controllers.OauthTokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/controllers.OauthTokenData"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.APIError"
                    }
                },
                "is_success": {
                    "type": "boolean"
                },
                "meta": {}
            }
        },
        "utils.APIError": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "string"
                },
                "error_messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}